move files and setup makefile
