+++ data-validation/Makefile
@@ -1,0 +1,21 @@
+# SPEC = Specification
+# TARGETS = $(SPEC).html $(SPEC).pdf
+# all: $(TARGETS)
+
+WETZEL = npx wetzel@0.2.2
+
+SCHEMALINK = schema
+EMBEDSCHEMA = JsonSchemaReference.adoc
+# Base name of the generated properties reference file
+PROPREF = PropertiesReference.adoc
+# Files generated by wetzel
+GENERATED = $(PROPREF) $(EMBEDSCHEMA)
+
+
+$(GENERATED): $(wildcard schema/*schema.json)
+	$(WETZEL) -n -a=cqo -m=a -p "$(SCHEMALINK)" -e "$(EMBEDSCHEMA)" \
+		-c "icon:check[]" -k "**MUST**"\
+		schema/root.schema.json > $(PROPREF)
+
+clean: 
+	-rm -f $(GENERATED) $(TARGETS)
