+++ data-validation/schema/feature-def.schema.json
@@ -1,0 +1,46 @@
+{
+    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$id": "featureDefSchema",
+    "title": "Feature Defs",
+    "description": "Measured features in the dataset",
+    "type": "object",
+    "additionalProperties": false,
+    "properties": {
+        "displayName": {
+            "description": "Human readable name",
+            "type": "string"
+        },
+        "description": {
+            "description": "Description of how the data was collected/measured",
+            "type": "string"
+        },
+        "tooltip": {
+            "description": "Shorter version of description",
+            "type": "string"
+        },
+        "unit": {
+            "description": "unit of measurement",
+            "type": "string"
+        },
+        "key": {
+            "description": "Id of the feature",
+            "type": "string"
+        },
+        "discrete": {
+            "description": "Whether it's a continuous measurement or not",
+            "type": "boolean"
+        },
+        "options": {
+            "description": "For discrete features, display items for each value",
+            "type": "object"
+        }
+    },
+    "required": [
+        "displayName",
+        "description",
+        "tooltip",
+        "unit",
+        "key",
+        "discrete"
+    ]
+}
