+++ data-validation/schema/measured-features.schema.json
@@ -1,0 +1,36 @@
+{
+    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$id": "measuredFeaturesSchema",
+    "title": "Measured features",
+    "description": "Data per cell needed to create the CFE plot",
+    "type": "array",
+    "items": {
+        "additionalProperties": false,
+        "type": "object",
+        "properties": {
+            "f": {
+                "description": "ordered array of measured features",
+                "type": "array",
+                "items": {
+                    "type": "number"
+                }
+            },
+            "p": {
+                "description": "Protein name",
+                "type": "string"
+            },
+            "t": {
+                "description": "Thumbnail Path",
+                "type": "string"
+            }
+
+
+        },
+        "required": [
+            "f",
+            "p",
+            "t"
+        ]
+    }
+
+}
