+++ data-validation/schema/root.schema.json
@@ -1,0 +1,22 @@
+{
+    "$schema": "http://json-schema.org/draft-04/schema",
+    "title": "Cell Feature Dataset Specification",
+    "type": "object",
+    "properties": {
+        "dataset": {
+            "$ref": "dataset.schema.json"
+        },
+        "feature-def": {
+            "$ref": "feature-def.schema.json"
+        },
+        "file-info": {
+            "$ref": "file-info.schema.json"
+        },
+        "manifest": {
+            "$ref": "manifest.schema.json"
+        },
+        "measured-features": {
+            "$ref": "measured-features.schema.json"
+        }
+    }
+}
